refactor(decompose): split packages/systems/graph-db-server/src/application/session/project.test.ts into folder
